Rename pyodide-bootstrap to python-entrypoint and add some comments
